+++ CHANGELOG.md
@@ -51,3 +51,9 @@
 - Fixed before and after texts issue when text is empty.
 - Updated README.md with online link and clarifications regarding HTTPS and local timer connection.
 - Added receiving faults and refusals from timer when FA connection is down.
+
+## 0.6.4 - 2024/09/29
+- Fixed image button not showing "Delete" when loaded from saved settings.
+- Added verification before deleting image.
+- Some spanish names in code translated to english
+- Improve code maintainability